fix(prototype/audio-proofread): split client-safe module out of server service
`clip-audio-proofread.ts` (no `.server` suffix) imported
`db-video-operations.server.ts`, so React Router couldn't tell it was
server-only and tried to bundle it for the client — failing the build
the moment the route's component referenced the runtime constant
DEFAULT_PROOFREAD_OPTIONS (React Router only strips server code out of
loader/action/middleware/headers exports, not the component itself).

Splits the file: types, tunable-option defaults/sanitization, and pure
helpers move to clip-audio-proofread-shared.ts (no server imports, safe
for the client route to import directly); the DB/ffmpeg-backed
ClipAudioProofreadService stays in clip-audio-proofread.server.ts,
which re-exports the shared module for existing server-side callers.

Verified `pnpm run build` (not just typecheck) now succeeds, plus
typecheck, lint:boundaries, and both test files still pass.
