+++ app/services/clip-audio-proofread-shared.ts
@@ -1,26 +1,19 @@
 /**
- * PROTOTYPE service backing `app/routes/prototype.audio-proofread.tsx`.
+ * Client-safe half of the audio-proofread prototype: types, tunable-option
+ * defaults, and pure helper functions with no server-only imports (no DB, no
+ * ffmpeg process spawning). Split out of `clip-audio-proofread.server.ts` so
+ * that `app/routes/prototype.audio-proofread.tsx` can reference
+ * `DEFAULT_PROOFREAD_OPTIONS` (a runtime value, not just a type) from its
+ * client-rendered component without dragging `db-video-operations.server.ts`
+ * into the browser bundle — React Router only strips server-only imports out
+ * of `loader`/`action`/`middleware`/`headers` exports, not out of the
+ * component itself. See https://reactrouter.com/explanation/code-splitting.
  *
- * Purpose: sanity-check whether ffmpeg silence-detection can flag the kind of
- * editing problems a human catches by ear in a rendered lesson — a pause that
- * runs long, a brief audio dropout, a click/level-jump right at a clip join —
- * so Matt can eyeball the results before anyone invests in a real feature.
- * No auto-fix, no publish-gating: this only produces a flat report.
- *
- * All the threshold constants below are rough first guesses, not tuned
- * against real footage — see the PR description for how to retune them.
- *
- * Deliberately ignores the render-time padding `course-publish-service.ts` /
- * `batch-export.server.ts` add (`LONG_PAUSE_DURATION` = 0.18s between a
- * `pauseType: "long"` clip and the next, `FINAL_VIDEO_PADDING` = 0.42s on the
- * very last clip). Matt's feedback timestamps are eyeballed by ear ("1:30"),
- * not frame-exact, so a few hundred ms of drift near a `long` pause or at the
- * very end of the video is acceptable noise for a throwaway prototype.
+ * `clip-audio-proofread.server.ts` re-exports everything here, plus the
+ * actual `ClipAudioProofreadService` (DB + ffmpeg), so server-side code can
+ * still import from a single place.
  */
 
-import { Effect } from "effect";
-import { FFmpegCommandsService } from "./ffmpeg-commands";
-import { VideoOperationsService } from "./db-video-operations.server";
 import { parseSilencePeriods } from "./silence-detection";
 import {
   SILENCE_THRESHOLD_DB,
@@ -119,13 +112,13 @@ export interface ProofreadClip {
 
 // ─── Pure helpers (unit-tested in clip-audio-proofread.test.ts) ────────────
 
-const round2 = (n: number) => Math.round(n * 100) / 100;
+export const round2 = (n: number) => Math.round(n * 100) / 100;
 
 /**
  * Walks the ordered, non-archived clip list and sums `sourceEndTime -
  * sourceStartTime` per clip to get each clip's start offset in the final
- * rendered (concatenated) video. Ignores render-time padding — see the
- * module doc comment.
+ * rendered (concatenated) video. Ignores render-time padding — see
+ * `clip-audio-proofread.server.ts`'s module doc comment.
  */
 export function computeClipVideoOffsets(
   clips: readonly ProofreadClip[]
@@ -278,7 +271,7 @@ export function findJoinHits(
  * short-cutout pass is actually a duplicate of a long pause (see
  * `excludeLongPeriods`). The thresholds fed to ffmpeg to produce `longRaw`
  * and `shortRaw` in the first place are the service's job, not this
- * function's — see `proofreadPerClipSpans` below.
+ * function's — see `proofreadPerClipSpans` in `clip-audio-proofread.server.ts`.
  */
 export function classifyPerClipSpans(
   longRaw: string,
@@ -324,161 +317,10 @@ export function classifyBoundarySpan(
   }));
 }
 
-function consecutivePairs<T>(items: readonly T[]): [T, T][] {
+export function consecutivePairs<T>(items: readonly T[]): [T, T][] {
   const pairs: [T, T][] = [];
   for (let i = 0; i < items.length - 1; i++) {
     pairs.push([items[i]!, items[i + 1]!]);
   }
   return pairs;
 }
-
-// ─── Service ─────────────────────────────────────────────────────────────
-
-export class ClipAudioProofreadService extends Effect.Service<ClipAudioProofreadService>()(
-  "ClipAudioProofreadService",
-  {
-    effect: Effect.gen(function* () {
-      const videoOps = yield* VideoOperationsService;
-      const ffmpeg = yield* FFmpegCommandsService;
-
-      const proofreadPerClipSpans = Effect.fn("proofreadPerClipSpans")(
-        function* (
-          clip: ProofreadClip,
-          videoStartSeconds: number,
-          options: ProofreadOptions
-        ) {
-          const [longRaw, shortRaw] = yield* Effect.all(
-            [
-              ffmpeg.detectSilence(clip.videoFilename, {
-                threshold: options.silenceThresholdDb,
-                silenceDuration: options.longPauseMinSeconds,
-                startTime: clip.sourceStartTime,
-              }),
-              ffmpeg.detectSilence(clip.videoFilename, {
-                threshold: options.silenceThresholdDb,
-                silenceDuration: options.shortCutoutMinSeconds,
-                startTime: clip.sourceStartTime,
-              }),
-            ],
-            { concurrency: 2 }
-          );
-
-          return classifyPerClipSpans(
-            longRaw,
-            shortRaw,
-            clip,
-            videoStartSeconds,
-            options
-          );
-        }
-      );
-
-      const proofreadBoundary = Effect.fn("proofreadBoundary")(function* (
-        clipA: ProofreadClip,
-        clipB: ProofreadClip,
-        clipBVideoStartSeconds: number,
-        options: ProofreadOptions
-      ) {
-        const { segmentA, segmentB, joinPointSeconds } = computeJoinWindow(
-          clipA,
-          clipB,
-          options.joinWindowSeconds
-        );
-
-        if (segmentA.durationSeconds <= 0 || segmentB.durationSeconds <= 0) {
-          return [];
-        }
-
-        const rawOutput = yield* ffmpeg.detectSilenceAcrossJoin(
-          {
-            file: segmentA.file,
-            startTime: segmentA.seekSeconds,
-            duration: segmentA.durationSeconds,
-          },
-          {
-            file: segmentB.file,
-            startTime: segmentB.seekSeconds,
-            duration: segmentB.durationSeconds,
-          },
-          {
-            threshold: options.silenceThresholdDb,
-            silenceDuration: options.shortCutoutMinSeconds,
-          }
-        );
-
-        return classifyBoundarySpan(
-          rawOutput,
-          joinPointSeconds,
-          clipB,
-          clipBVideoStartSeconds,
-          options.joinToleranceSeconds
-        );
-      });
-
-      const proofreadVideo = Effect.fn("proofreadVideo")(function* (
-        videoId: string,
-        overrides?: Partial<Record<keyof ProofreadOptions, unknown>> | null
-      ) {
-        const options = sanitizeProofreadOptions(overrides);
-
-        const video = yield* videoOps.getVideoWithClipsById(videoId);
-        const clips: ProofreadClip[] = video.clips.map((c) => ({
-          id: c.id,
-          videoFilename: c.videoFilename,
-          sourceStartTime: c.sourceStartTime,
-          sourceEndTime: c.sourceEndTime,
-        }));
-
-        const offsets = computeClipVideoOffsets(clips);
-        const videoStartByClipId = new Map(
-          offsets.map((o) => [o.clipId, o.videoStartSeconds])
-        );
-        const totalDurationSeconds = offsets.reduce(
-          (acc, o) => acc + o.durationSeconds,
-          0
-        );
-
-        const perClipSpans = yield* Effect.forEach(
-          clips,
-          (clip) =>
-            proofreadPerClipSpans(
-              clip,
-              videoStartByClipId.get(clip.id)!,
-              options
-            ),
-          { concurrency: 4 }
-        );
-
-        const boundarySpans = yield* Effect.forEach(
-          consecutivePairs(clips),
-          ([clipA, clipB]) =>
-            proofreadBoundary(
-              clipA,
-              clipB,
-              videoStartByClipId.get(clipB.id)!,
-              options
-            ),
-          { concurrency: 4 }
-        );
-
-        const spans = [...perClipSpans.flat(), ...boundarySpans.flat()].sort(
-          (a, b) => a.videoTimestampSeconds - b.videoTimestampSeconds
-        );
-
-        return {
-          videoId: video.id,
-          title: video.title,
-          totalDurationSeconds: round2(totalDurationSeconds),
-          options,
-          spans,
-        };
-      });
-
-      return { proofreadVideo };
-    }),
-    dependencies: [
-      VideoOperationsService.Default,
-      FFmpegCommandsService.Default,
-    ],
-  }
-) {}
